Functions work based on roles

+++ app/src/main/java/ktu/edu/projektas/app/ui/schedule/ScheduleFragment.kt
@@ -25,6 +25,7 @@ import ktu.edu.projektas.app.data.Event
 import ktu.edu.projektas.app.ui.schedule.ScheduleAdapter
 import androidx.appcompat.widget.AppCompatButton
 import com.google.android.material.textfield.TextInputEditText
+import com.google.firebase.auth.FirebaseAuth
 import ktu.edu.projektas.databinding.FragmentScheduleBinding
 
 
@@ -34,6 +35,7 @@ class ScheduleFragment : Fragment() {
 
     private var semesterStart : Long? = null
     private var semesterEnd : Long? = null
+    private val user = FirebaseAuth.getInstance().currentUser
 
     private lateinit var spinner : Spinner
     private val adapter =   ScheduleAdapter(clickListener = this::onLongClick,
@@ -129,6 +131,7 @@ class ScheduleFragment : Fragment() {
     ): View? {
         binding = FragmentScheduleBinding.inflate(inflater)
 
+
         viewModel.events.observe(viewLifecycleOwner){
             adapter.submitList(it)
         }
@@ -156,15 +159,21 @@ class ScheduleFragment : Fragment() {
         return Instant.ofEpochMilli(value!!).atZone(ZoneId.systemDefault()).toLocalDate()
     }
     private fun onLongClick(event: Event) {
-        AlertDialog.Builder(context)
+        if(event.userUUID == user!!.uid) {
+            AlertDialog.Builder(context)
                 .setTitle("Delete entry")
                 .setMessage("Are you sure you want to delete this entry?")
                 .setPositiveButton(android.R.string.yes) { dialog, which ->
-                    viewModel.deleteByGroup(event.groupId)
+                    if (event.groupId == 0) {
+                        viewModel.deleteByGroup(event.groupId)
+                    } else {
+                        viewModel.deleteByFirebaseId(event.firebaseId)
+                    }
                 }
                 .setNegativeButton(android.R.string.no, null)
                 .setIcon(android.R.drawable.ic_dialog_alert)
                 .show()
+        }
     }
 
     private fun onClick(event: Event) {
